+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This version introduces breaking changes. Existing code might better stick to [v
 ### Changed
 - `global_reuse`, `instance_reuse`, `reopen_variable_scope`, `root_variable_scope` and `VarScopeObject` have been rewritten, and their behaviors have been slightly changed.
   This might cause existing code to be malfunction, if these code relies heavily on the precise variable scope or name scope of certain variables or tensors.
+- Constructing a `VarScopeObject` within a method or a function decorated by `global_reuse` or `instance_reuse` has been totally disallowed.
 - `Trainer` now accepts `summaries` argument on construction.
 
 ### Removed